Avoiding ciclic dependency between packages
