+++ vraptor-core/src/main/java/br/com/caelum/vraptor4/http/route/DefaultRouter.java
@@ -39,6 +39,7 @@ import br.com.caelum.vraptor4.http.ParameterNameProvider;
 import br.com.caelum.vraptor4.proxy.Proxifier;
 import br.com.caelum.vraptor4.util.collections.Filters;
 
+import com.google.common.base.Predicate;
 import com.google.common.collect.Collections2;
 import com.google.common.collect.Iterators;
 
@@ -115,7 +116,7 @@ public class DefaultRouter implements Router {
 
 
 	private Collection<Route> routesMatchingUriAndMethod(String uri, HttpMethod method) {
-		Collection<Route> routesMatchingMethod = Collections2.filter(routesMatchingUri(uri), Filters.allow(method));
+		Collection<Route> routesMatchingMethod = Collections2.filter(routesMatchingUri(uri), allow(method));
 		if (routesMatchingMethod.isEmpty()) {
 			EnumSet<HttpMethod> allowed = allowedMethodsFor(uri);
 			throw new MethodNotAllowedException(allowed, method.toString());
@@ -133,7 +134,7 @@ public class DefaultRouter implements Router {
 	}
 
 	private Collection<Route> routesMatchingUri(String uri) {
-		Collection<Route> routesMatchingURI = Collections2.filter(routes, Filters.canHandle(uri));
+		Collection<Route> routesMatchingURI = Collections2.filter(routes, canHandle(uri));
 		if (routesMatchingURI.isEmpty()) {
 			throw new ControllerNotFoundException();
 		}
@@ -147,7 +148,7 @@ public class DefaultRouter implements Router {
 
 		if (proxifier.isProxyType(type)) rawtype = type.getSuperclass();
 
-		Iterator<Route> matches = Iterators.filter(routes.iterator(), Filters.canHandle(rawtype, method));
+		Iterator<Route> matches = Iterators.filter(routes.iterator(), canHandle(rawtype, method));
 		if (matches.hasNext()) {
 			try {
 				return matches.next().urlFor(rawtype, method, params);
@@ -165,4 +166,27 @@ public class DefaultRouter implements Router {
 		return Collections.unmodifiableList(new ArrayList<Route>(routes));
 	}
 
+    private Predicate<Route> canHandle(final Class<?> type, final Method method) {
+        return new Predicate<Route>() {
+            public boolean apply(Route route) {
+                return route.canHandle(type, method);
+            }
+        };
+    }
+
+    private Predicate<Route> canHandle(final String uri) {
+        return new Predicate<Route>() {
+            public boolean apply(Route route) {
+                return route.canHandle(uri);
+            }
+        };
+    }
+    
+    private Predicate<Route> allow(final HttpMethod method) {
+        return new Predicate<Route>() {
+            public boolean apply(Route route) {
+                return route.allowedMethods().contains(method);
+            }
+        };
+    }
 }
